src/blocks: unravel type shenanigans from template
Removing the `ts-expect-error` from the `RenderBlocks` component
uncovered a couple of type errors in how we were passing data - notably,
the `Form` component was not correctly honoring the case where the
passed-in `Form` was just an ID, not a whole object.

If we want to tranform the generated types into types that do not allow
their member data to have objects-as-references, we will need to do that
at the `page.tsx` level as getting at the `payload` object and using it
to fetch data is an async process, which we can't do *inside* of a JSX
element constructor.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>